+++ src/test/test_time.cpp
@@ -12,7 +12,7 @@ TEST(dmTime, GetTime)
     uint64_t start = dmTime::GetTime();
     dmTime::Sleep(200000);
     uint64_t end = dmTime::GetTime();
-    ASSERT_NEAR(200000, end-start, 20000);
+    ASSERT_NEAR((double) 200000, (double) (end-start), (double) 20000);
 }
 
 int main(int argc, char **argv)
